backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
playwright==1.47.0
weasyprint>=68.1
aiosmtplib==3.0.1
jinja2==3.1.4
pyyaml==6.0.2
apscheduler==3.10.4
httpx==0.27.2
